Specify type as float

+++ hendrics/lcurve.py
@@ -422,7 +422,7 @@ def lcurve_from_events(
 
     n_times = int((tstop - tstart) / bintime)
     ev_times = (events - tstart).astype(float)
-    time_ranges = [0, n_times * bintime]
+    time_ranges = [0, float(n_times * bintime)]
     time_edges = np.linspace(time_ranges[0], time_ranges[1], n_times + 1).astype(float)
 
     raw_counts = histogram(
